Fixed UFNs of non-dissected 2015-16 fish in fish_field_data
Automatically-generated UFNs did not follow the order of SEMSP IDs due to incorrect sorting (alphabetical instead of natural). Now, U5501-U9817 and U18001-U19872 follow the relative order of SEMSP IDs for each species in the seine (i.e. the order in which they were taken out of the net).